Detect empty file list in commit and handle accordingly
Should a git commit contain no file changes, a situation which is
possible when the git repository has been converted from another SCM
system such as Subversion, then assign an empty string to the list of
files.

This will prevent an exception caused when ruby tries to split a Nil
object.

Tested against git 1.8.0.1 and ruby 1.9.3